* fixed clang warnings

+++ iOS/HOWTO/SimpleSubscribe/3.4/PubNubDemo/libs/PubNub/Network/PNReachability.m
@@ -165,7 +165,7 @@ PNReachabilityStatus PNReachabilityStatusForFlags(SCNetworkReachabilityFlags fla
 
                 status = PNReachabilityStatusReachableViaWiFi;
 
-                unsigned int flagsForCleanUp = (unsigned int)flags;
+                unsigned long flagsForCleanUp = (unsigned long)flags;
                 PNBitsOff(&flagsForCleanUp, kSCNetworkReachabilityFlagsReachable, kSCNetworkReachabilityFlagsIsDirect,
                                             kSCNetworkReachabilityFlagsIsLocalAddress, BITS_LIST_TERMINATOR);
                 flags = (SCNetworkReachabilityFlags)flagsForCleanUp;
